Refactor stop related hooks to their own directory
