+++ src/hooks/index.ts
@@ -2,20 +2,17 @@ export * from './redux';
 export * from './search/useSearch';
 export * from './search/useSearchQueryParser';
 export * from './search/useSearchResults';
+export * from './stops';
 export * from './useAsyncQuery';
 export * from './useCheckValidityAndPriorityConflicts';
 export * from './useContextStateSelector';
 export * from './useCreateLine';
-export * from './useCreateStop';
 export * from './useDeleteRoute';
-export * from './useDeleteStop';
 export * from './useEditLine';
 export * from './useEditRoute';
 export * from './useEditRouteGeometry';
-export * from './useEditStop';
 export * from './useExtractRouteFromFeature';
 export * from './useFilterStops';
 export * from './useGetDisplayedRoutes';
-export * from './useGetStopLinkAndDirection';
 export * from './useShowRoutesOnModal';
 export * from './useUrlQuery';
